refactoring: Constants moved to opba-facade-protocol-api-shared
