--- opba-banking-protocol-facade/src/main/java/de/adorsys/opba/protocol/facade/util/logresolver/Constants.java
@@ -1,19 +1,0 @@
-package de.adorsys.opba.protocol.facade.util.logresolver;
-
-import lombok.experimental.UtilityClass;
-
-
-/**
- * LogResolver constant names.
- */
-
-
-@UtilityClass
-@SuppressWarnings("checkstyle:HideUtilityClassConstructor") // Lombok generates private ctor.
-public class Constants {
-
-    /**
-     * Commonly used keyword for Log-classes.
-     */
-    public static final String NULL = "null";
-}
