+++ opba-banking-protocol-facade/src/main/java/de/adorsys/opba/protocol/facade/util/logresolver/domain/request/RequestLog.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 
-import static de.adorsys.opba.protocol.facade.util.logresolver.Constants.NULL;
+import static de.adorsys.opba.protocol.api.Constants.NULL;
 
 
 @Getter
